Домашняя работа 3. Реализованы формы и переходы по нажатию на постер в первом окне и при нажатии на кнопку "back" во втором окне
