--- app/src/main/res/drawable/custom_ratingbar_filled.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:drawable="@drawable/staricon_select" android:state_pressed="true" android:state_window_focused="true"/>
-    <item android:drawable="@drawable/staricon_select" android:state_focused="true" android:state_window_focused="true"/>
-    <item android:drawable="@drawable/staricon_select" android:state_selected="true" android:state_window_focused="true"/>
-    <item android:drawable="@drawable/staricon_select"/>
-</selector>
