+++ app/src/main/res/drawable/custom_ratingbar_small.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android" >
+
+    <item
+        android:id="@+id/background"
+        android:drawable="@drawable/staricon_small_unselect"/>
+    <item
+        android:id="@+id/secondaryProgress"
+        android:drawable="@drawable/staricon_small_unselect"/>
+    <item
+        android:id="@+id/progress"
+        android:drawable="@drawable/staricon_small_select"/>
+
+</layer-list>
