+++ app/src/main/res/drawable/movie_gradient_roundrect.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle" >
+
+    <!--<corners android:radius="8dp" />-->
+
+    <!--<gradient
+        android:angle="90"
+        android:endColor="@color/main"
+        android:startColor="@color/main"
+        android:type="linear" />-->
+    <gradient
+        android:angle="90"
+        android:endColor="@color/oogray"
+        android:startColor="@color/ffgray"
+        android:centerX="15%"
+        android:centerColor="@color/ffgray" />
+
+    <corners android:radius="4dp" />
+
+
+    <!--<stroke
+        android:width="1dp"
+        android:color="#3B3B4D, 50 %" />-->
+
+    <!--<padding
+        android:bottom="15dp"
+        android:left="15dp"
+        android:right="15dp"
+        android:top="15dp" />-->
+
+</shape>
