+++ app/src/main/res/drawable/movie_roundrect.xml
@@ -2,7 +2,7 @@
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle" >
 
-    <corners android:radius="15dp" />
+    <corners android:radius="8dp" />
 
     <!--<gradient
         android:angle="90"
@@ -12,7 +12,7 @@
 
     <stroke
         android:width="1dp"
-        android:color="#FF000000" />
+        android:color="@color/movie_rectangle_color" />
 
     <!--<padding
         android:bottom="15dp"
