+++ app/src/main/res/drawable/movie_roundrect_age.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle" >
+
+    <corners android:radius="5dp" />
+
+    <!--<gradient
+        android:angle="90"
+        android:endColor="#FFFFFFFF"
+        android:startColor="#FF000000"
+        android:type="linear" />-->
+
+    <stroke
+        android:width="1dp"
+        android:color="@color/main" />
+    <solid android:color="@color/main" />
+
+    <!--<padding
+        android:bottom="15dp"
+        android:left="15dp"
+        android:right="15dp"
+        android:top="15dp" />-->
+
+</shape>
