+++ app/src/main/res/layout/activity_movie_details.xml
@@ -133,6 +133,7 @@
             android:text="@string/storyline"
             android:textColor="@color/white"
             app:layout_constraintEnd_toStartOf="@+id/main_indent_end_guideline"
+            app:layout_constraintHorizontal_bias="1.0"
             app:layout_constraintStart_toStartOf="@+id/main_indent_start_guideline"
             app:layout_constraintTop_toBottomOf="@+id/movie_posts_text_view" />
