+++ app/src/main/res/values/strings.xml
@@ -11,6 +11,21 @@
     <string name="movie_name">Avengers:  End Game</string>
     <string name="movie_ganr">Action, Adventure, Fantasy</string>
     <string name="movie_reviews">125 Reviews</string>
+    <string name="movie_2_age">16+</string>
+    <string name="movie_2_name">Tenet</string>
+    <string name="movie_2_ganre">Action, Sci-Fi, Thriller</string>
+    <string name="movie_2_reviews">98 Reviews</string>
+    <string name="movie_2_duration_text">97 min</string>
+    <string name="movie_3_age">13+</string>
+    <string name="movie_3_name">Black Widow</string>
+    <string name="movie_3_ganre">Action, Adventure, Sci-Fi</string>
+    <string name="movie_3_reviews">38 Reviews</string>
+    <string name="movie_3_duration_text">102 min</string>
+    <string name="movie_4_age">13+</string>
+    <string name="movie_4_name">Wonder Woman 1984</string>
+    <string name="movie_4_ganre">Action, Adventure, Fantasy</string>
+    <string name="movie_4_reviews">74 Reviews</string>
+    <string name="movie_4_duration_text">120 min</string>
     <string name="storyline">Storyline</string>
     <string name="storyline_text">After the devastating events of Avengers: Infinity War, the universe is in ruins. With the help of remaining allies, the Avengers assemble once more in order to reverse Thanos actions and restore balance to the universe.</string>
     <string name="cast">Cast</string>
@@ -25,4 +40,6 @@
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="name_movies_list">Movies list</string>
+    <string name="movie_duration_1_text">137 min</string>
+    <string name="contentDescription">1:1</string>
 </resources>
