+++ app/src/main/res/values/stylebar_small.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+        <style name="CustomRatingBar_Small" parent="@android:style/Widget.RatingBar">
+            <item name="android:progressDrawable">@drawable/staricon_small_select</item>
+            <item name="android:minHeight">8dp</item>
+            <item name="android:maxHeight">8dp</item>
+        </style>
+</resources>
